Move capability instances to dedicated objects
This makes them orphan, but we do not want to have the responsibility
of making non-orphan instances for types we don't control, when the
capabilities form a subclass hierarchy.
