+++ modules/core/src/smithy4s/capability/Zipper.scala
@@ -16,9 +16,6 @@
 
 package smithy4s.capability
 
-import scala.concurrent.Future
-import scala.concurrent.ExecutionContext
-
 /**
   * A zipper is similar in notion to `Applicative`, in that it allows for
   * combining several values of `F` in non-sequential ways.
@@ -44,44 +41,5 @@ trait Zipper[F[_]] extends Covariant[F] {
 }
 
 object Zipper {
-
   def apply[F[_]](implicit ev: Zipper[F]): ev.type = ev
-
-  implicit def eitherZipper[E]: Zipper[Either[E, *]] =
-    new Zipper[Either[E, *]] {
-      def pure[A](a: A): Either[E, A] = Right(a)
-
-      override def zipMap[A, B, C](fa: Either[E, A], fb: Either[E, B])(
-          f: (A, B) => C
-      ): Either[E, C] = (fa, fb) match {
-        case (l @ Left(_), _)     => l.asInstanceOf[Either[E, C]]
-        case (_, r @ Left(_))     => r.asInstanceOf[Either[E, C]]
-        case (Right(a), Right(b)) => Right(f(a, b))
-      }
-
-      override def zipMapAll[A](
-          seq: IndexedSeq[Either[E, Any]]
-      )(f: IndexedSeq[Any] => A): Either[E, A] = {
-        val builder = IndexedSeq.newBuilder[Any]
-        var i = 0
-        var error: Left[E, Any] = null
-        while (error == null && i < seq.size) {
-          seq(i) match {
-            case l @ Left(_) => error = l.asInstanceOf[Left[E, Any]]
-            case Right(r)    => builder += r
-          }
-          i += 1
-        }
-        if (error != null) error.asInstanceOf[Left[E, A]]
-        else Right(f(builder.result()))
-      }
-    }
-
-  implicit def futureZipper(implicit ec: ExecutionContext): Zipper[Future] =
-    new Zipper[Future] {
-      override def zipMapAll[A](seq: IndexedSeq[Future[Any]])(
-          f: IndexedSeq[Any] => A
-      ): Future[A] = Future.sequence(seq).map(f)
-      def pure[A](a: A): Future[A] = Future.successful(a)
-    }
 }
