+++ modules/test-utils/src/smithy4s/tests/JsonProtocolF.scala
@@ -23,6 +23,7 @@ import cats.syntax.all._
 import cats.MonadThrow
 import smithy4s.kinds._
 import smithy4s.codecs._
+import smithy4s.capability.instances.either._
 
 /**
   * These are toy interpreters that turn services into json-in/json-out
